AI-chat-builder/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
groq==0.4.1
google-generativeai==0.3.2
together==0.2.7
sentence-transformers==2.3.1
pypdf2==3.0.1
python-docx==1.1.0
cryptography==42.0.0
httpx==0.26.0
numpy==1.26.4
